Lots of refactoring to separate the demo application from the bflib framework.  befairController is now named as betfairController.class.php and is now the controller class for the bflib library. betfairController is totally unaware of the demo app.  betfairDemoRequestHandler.php has become the controlling code for the demo app (call that the demo application controller if you like).
The betfairView.class.php has been renamed bflibDemoView.class.php, although this will probably revert back to betfiarDemoView.class.php to align naming conventions with
the application controller. Again, this is not part of the core bflib library.

Ultimately, those classes which are part of the bflib core library, should be named as such, so betfairDialogue.class.php should become bflibDialogue.class.php and so
on.